7699 refactor mixins
