--- netbox/virtualization/constants.py
@@ -1,4 +1,0 @@
-# models values for ContentTypes which may be Cluster scope types
-CLUSTER_SCOPE_TYPES = (
-    'region', 'sitegroup', 'site', 'location',
-)
